pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-genomics"
version = "0.1.0"
description = "Agentic AI systems for genomics — starting with an autonomous variant interpretation copilot."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Ankur Sharma", email = "learning.ankur.ai@gmail.com" },
]
keywords = [
    "genomics",
    "bioinformatics",
    "agentic-ai",
    "llm",
    "langgraph",
    "variant-interpretation",
    "clinical-genomics",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langgraph>=0.2.50",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "anthropic>=0.39.0",
    "pydantic>=2.8",
    "pysam>=0.22.1",
    "requests>=2.32",
    "httpx>=0.27",
    "tenacity>=9.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "typer>=0.12",
    "pandas>=2.2",
]

[project.optional-dependencies]
demo = [
    "streamlit>=1.38",
    "jupyter>=1.0",
    "matplotlib>=3.9",
    "plotly>=5.24",
]
openai = [
    "langchain-openai>=0.2.0",
]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.urls]
Homepage = "https://github.com/ankurgenomics/agentic-genomics"
Repository = "https://github.com/ankurgenomics/agentic-genomics"
Issues = "https://github.com/ankurgenomics/agentic-genomics/issues"

[project.scripts]
agentic-genomics = "agentic_genomics.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_genomics"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "network: tests that require network access (skipped by default in CI without keys)",
    "slow: tests that are slow",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
